调整电流相关系数以提高扭矩控制精度
- 将 Axis::get_axis_state 函数中的实际扭矩计算从 int16_t 更改为 float 类型
- 在 Axis 和 Encoder 类中更新电流相关系数，提高分辨率和控制精度
- 调整 Encoder 类中的校准范围和带宽设置，优化性能

+++ Firmware/MotorControl/axis.cpp
@@ -87,7 +87,7 @@ static void step_cb_wrapper(void* ctx) {
 void Axis::get_axis_state(axis_state_t* state)
 {
 
-    int16_t actual_torque = motor_.convert_torque_from_current(motor_.current_control_.Iq_measured, motor_.config_.CURRENT2TORQUE_COEFF, motor_.NUM_LINEARITY_SEG,  motor_.CALIBRATION_INCREMENT);
+    float actual_torque = motor_.convert_torque_from_current(motor_.current_control_.Iq_measured, motor_.config_.CURRENT2TORQUE_COEFF, motor_.NUM_LINEARITY_SEG,  motor_.CALIBRATION_INCREMENT);
     state->erro =  axis_state_.erro;
     state->pos = saturation((int32_t)(encoder_.gearboxpos_ * position_coeff_motor2encos +32768),0,65535 );   // 2pi*12.5*32768
     if(config_.gear_vel_used == true)
